+++ pkg/upstream/http.go
@@ -2,13 +2,13 @@ package upstream
 
 import (
 	"crypto/tls"
-	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/middleware"
 	"net/http"
 	"net/http/httputil"
 	"net/url"
 	"strings"
 
 	"github.com/mbland/hmacauth"
+	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/apis/middleware"
 	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/apis/options"
 	"github.com/yhat/wsutil"
 )
@@ -78,7 +78,9 @@ type httpUpstreamProxy struct {
 // request headers
 func (h *httpUpstreamProxy) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 	scope := middleware.GetRequestScope(req)
-	scope.Upstream = h.upstream
+	if scope != nil {
+		scope.Upstream = h.upstream
+	}
 
 	// TODO (@NickMeves) - Deprecate GAP-Signature & remove GAP-Auth
 	if h.auth != nil {