+++ pkg/upstream/static.go
@@ -4,8 +4,8 @@ import (
 	"fmt"
 	"net/http"
 
+	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/apis/middleware"
 	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/logger"
-	"github.com/oauth2-proxy/oauth2-proxy/v7/pkg/middleware"
 )
 
 const defaultStaticResponseCode = 200
@@ -28,7 +28,9 @@ type staticResponseHandler struct {
 // ServeHTTP serves a static response.
 func (s *staticResponseHandler) ServeHTTP(rw http.ResponseWriter, req *http.Request) {
 	scope := middleware.GetRequestScope(req)
-	scope.Upstream = s.upstream
+	if scope != nil {
+		scope.Upstream = s.upstream
+	}
 
 	rw.WriteHeader(s.code)
 	_, err := fmt.Fprintf(rw, "Authenticated")